CommandBuilder Harness + Tests (#24)
* write initial tests for CommandBuilder inputs

* add static output test

* add dynamic test, and failing rawcall test

* harness completed and working with a basic set of input/output test combinations

* fix some info messages

* prettify

* remove stray debug line

* update npm package shit

* readd strings from main

* update type in strcat test to match main branch Strings.sol

* change according to feedback from nick

* Update CommandBuilderHarness.sol

* prettify

+++ contracts/test/StateTest.sol
@@ -1,8 +1,16 @@
 pragma solidity ^0.8.4;
 
 contract StateTest {
-    function addSlots(uint dest, uint src, uint src2, bytes[] memory state) public pure returns(bytes[] memory) {
-        state[dest] = abi.encode(abi.decode(state[src], (uint)) + abi.decode(state[src2], (uint)));
+    function addSlots(
+        uint256 dest,
+        uint256 src,
+        uint256 src2,
+        bytes[] memory state
+    ) public pure returns (bytes[] memory) {
+        state[dest] = abi.encode(
+            abi.decode(state[src], (uint256)) +
+                abi.decode(state[src2], (uint256))
+        );
         return state;
     }
 }
